requirements.txt
Django==3.1.4
djangorestframework==3.11.2
djangorestframework-simplejwt==4.7.0
mysqlclient==2.0.1